// src/app/Life/Workflow/service/workflow.interface.ts
export interface IWorkflow {
  timeToSla: any;
  id: number;
  wfNo: string;
  code: string;
  description: string;
  assignedBy: string;
  assignedTo: string;
  assignedToGroup: string;
  clientNo: string;
  narration: string;
  createdOn: ITiming;
  updatedOn: null;
  createdBy: null;
  updatedBy: null;
  originalSla: number;
  revisedSla: number;
  revisedSlaUser: string;
  dueOn: null;
  refEmailAddr: string;
  externalRef: string;
  finishOn: ITiming;
  nextTaskDelayTime: number;
  maintainFunction: string;
  modifiedOn: ITiming;
  modifiedUser: string;
  predecessor: string;
  successor: string;
  schedule: boolean;
  source: string;
  startOn: ITiming;
  status: string;
  scheduleNo: string;
  busLine: string;
  wfCompScore: number;
  wfOsSlaScore: number;
  wfEscScore: number;
  wfErrScore: number;
  wfScore: number;
  category: null;
  title: null;
  branch: null;
  sensitivity: null;
  boxNo: null;
  timeOsSla: number;
  colourCode: string;
  escalation: boolean;
  errorSeverity: string;
  slaLevel: string;
  error: boolean;
  active: null;
  errorReason: string;
  errorLevel: string;
  errorNarration: string;
  companyCode: string;
  refCat: string;
  refNo: string;
  linkedWorkflow: string;
}

export interface ITiming {
  month: string;
  dayOfWeek: string;
  dayOfYear: number;
  nano: number;
  year: number;
  monthValue: number;
  dayOfMonth: number;
  hour: number;
  minute: number;
  second: number;
  chronology: {
    calendarType: string;
    id: string;
  };
}

export interface IWorkflowDocument {
  agent: string;
  authorizedUser: string;
  availToPh: boolean;
  branch: string;
  busLine: string;
  clientNo: string;
  company: string;
  createdDate: Date;
  createdTime: number;
  createdUser: string;
  crmClientNo: string;
  dispatched: boolean;
  dispatchedDate: Date;
  dispatchedTime: number;
  dnNo: string;
  docBox: string;
  docCat: string;
  docId: string;
  docSubCat: string;
  documentContent: string;
  documentKey: number;
  documentSource: string;
  emailForm: string;
  emailFrom: string;
  event: string;
  exportNo: number;
  externalRef: string;
  fileName: string;
  groupNo: string;
  groupQuoteNo: string;
  hidden: boolean;
  how: string;
  memberNo: string;
  originalSla: number;
  policyCode: string;
  policyNo: string;
  quoteNo: string;
  refCat: string;
  refNo: string;
  revisedSla: number;
  saved: boolean;
  senderId: string;
  sensitivity: string;
  spoolFolder: string;
  statement: string;
  code: string;
  title: string;
  upload: string;
  wfNo: string;
}

export interface ISnoozeInfo {
  auth: boolean;
  id: number;
  snoozeAuthBy: string;
  snoozeBasis: string;
  snoozeBy: string;
  snoozeComment: string;
  snoozeDate: string;
  snoozeReason: string;
  snoozeTime: number;
  snoozeUntilDate: Date;
  snoozeUntilTime: number;
  workflowSnoozeHistoryId: number;
}
export interface IWorkflowSchedule {
  wfNo: any;
  id: number;
  scheduleFreq: string;
  scheduleSla?: any;
  lastScheduleOn: string;
  nextScheduleOn: string;
  lastScheduleCompleteOn: string;
  code: string;
  scheduleAssignedTo: string;
  scheduleAssignedToGroup: string;
  scheduleSlaDays: number;
  scheduleSlaHours: number;
  description: string;
  active: boolean;
  refCat: string;
  refNo: string;
  category?: any;
  title?: any;
  branch?: any;
  sensitivity?: any;
  boxNo?: any;
  fileName: string;
}